Move intoOrders into useOrders hook
